Add build_legacy_v1_feeds.py
This script is in place to provide a "python" way in building the v1 feed. This is based off the original `build-sofa-feed.py` script which serves well in self hosted and community feed processin eg. heavily used in SOFA 1.0.

This script must be used in a GH action, take note of USAGE &HARD REQUIREMENTS in the docstring